// models/Personnels.ts
/**
 * Created by S1-DEV-ETC on 20/03/2018.
 */

export interface Personnels{
  id:number;
  idE:number;
  cin: number;
  nom: string;
  prenom: string;
  civilite: string;
  signiature: string;
  genre: string;
  penseBete: string;
  idCU: number;
  adressePostale: string;
  fonction: string;
  email: string;
  creePar: number;
  modifiePar: number;
  createdAt:string;
  updatedAt:string;
  tel1:number;
  tel2:number;
  complement_adresse:string;
  ville:string;
  code_postale:string;
  avatar:string;
}